fix: Updated party room reservation ragistration
