+++ togather-common/src/main/java/com/togather/member/service/MemberService.java
@@ -61,7 +61,7 @@ public class MemberService {
     }
 
     @Transactional
-    public void withdrawal(Long memberSrl) {
+    public void delete(Long memberSrl) {
         //TODO: 비밀번호 검증 로직 추가할 가능성 있음!!
 
         Member findMember = memberRepository.findById(memberSrl)
@@ -72,10 +72,9 @@ public class MemberService {
         log.info("delete member: {}", findMember.getMemberSrl());
     }
 
-    @Transactional
-    public Member findById(long id) {
-        return memberRepository.findById(id)
-                .orElseThrow(RuntimeException::new);
+    public MemberDto findMemberDtoById(long id) {
+        return memberConverter.convertToDto(memberRepository.findById(id)
+                .orElseThrow(RuntimeException::new));
     }
 
 }
